Changed so that RoomSize is stored as a byte.

+++ LibraryAndService/Models/Room.cs
@@ -12,10 +12,10 @@ namespace LibraryAndService.Models
         public decimal Price { get; set; }
         public RoomType RoomType { get; set; }
         [Range(0, 100)]
-        public int RoomSize { get; set; }
+        public byte RoomSize { get; set; }
         public bool IsActive { get; set; }
 
-        public Room(string roomName, decimal price, RoomType roomType, int roomSize, bool isActive)
+        public Room(string roomName, decimal price, RoomType roomType, byte roomSize, bool isActive)
         {
             RoomName = roomName;
             Price = price;
